requirements.txt
pydantic
python-dotenv
langgraph
langchain-core
langchain-deepseek
langchain-tavily
langsmith
langchain-openai
langchain-text-splitters
langchain-community
faiss-cpu
langgraph_supervisor
matplotlib
seaborn
pandas
pymysql
scikit-learn
langgraph-cli[inmem]
